Separated test cases by component, for now, only for the required pages
